pyproject.toml
[project]
name = "events-face"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2.9,<6.0.0",
    "django-environ>=0.12.0,<0.13.0",
    "django-filter>=25.2,<26.0",
    "djangorestframework>=3.16.1,<4.0.0",
    "djangorestframework-simplejwt>=5.5.1,<6.0.0",
    "gunicorn>=23.0.0,<24.0.0",
    "requests>=2.32.5,<3.0.0",
]

[dependency-groups]
dev = ["ruff>=0.14.7,<0.15.0"]
docs = []

[tool.uv]
add-bounds = "major"

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
